[CMPLRLLVM-52569] Always parse Boolean constants as unsigned to avoid incorrect sign extension of the values. (#17673)
When IV comes from phi STMT of the form iv = phi [ true, %entry] [false,
%loop], the final loop is formed as
DO i1 0,1,1 {...} . The IV is parsed as (-1 *i1 + 1) which makes it
'true' when i1==0 and 'false' when i1==1. Previously we considered all
constants as signed that resulted in 'true' being parsed as '-1' due to
sign extension on 'i1 1' value. Now i1 values are considered unsigned by
default.
CMPLRLLVM-52569