[CMPLRLLVM-53241] Bugfix for loop fusion due to merging unexpected nodes caused by cycles and/or multiple pathways in fusegraph (#17654)
CMPLRLLVM-53241

The root cause for the bug is due to the fusion algorithm trying to
merge all nodes that have a path from A to B. However, in presence of
multiple paths, we were not checking to see if the path through the
intermediate nodes were a "good set". The fix was to add those checks
during the graph traversal to determine if the intermediate node is part
of a good path.

2 bugs came up. For debug compiler, we were trying to refine DD Edges
for backward and forward edges. However, it caused the dump() function
to try to print a null object. Secondly, the new testcase had a cycle in
the fusegraph. During sorting, compiler crashed due to the cycle causing
the BFS algorithm to infinitely loop during the traversal. Added a visit
flag during BFS to disable cycling.