[VPlan][CallVecDecisions] Prefer vectorizing a call with vector-library instead of vector-variants. (#17630)
Based on the current sequence of checks in VPlanCallVecDecisions, we try
to vectorize a call using its matching vector-variant even if it has an
appropriate match in vector-library. This relaxes this order by
preferring vector-library based matches over vector-variants. This is
done for better performance and more FP consistency in different
versions of outgoing loops generated by VPlan.

JIRA: CMPLRLLVM-53171