[VPlan][VSLP] Add code to recognize mixed operations SLP vectorization. (#17753)
The change adds the code, which is able to determine if there are two
differently vectorizable parts in input array. Currently only different
opcodes are supported in this context. Such, if input array contains
'+' and '-' operations that are vectorizable, the code takes costs of
full vector '+' and '-' and the cost of blending those two vectors to
form the requested vector.

The algorithm tries to vectorize elements remaining in input array after
the first try. If it succeeds, than the cost is of second vector accumulated
altogether with the cost of results blending operation.

The elements that appear to form the second vector are inserted into the
the first vectorization try vector in the order they met in input array.
With that done SLP graph building code continues normally.