[VPlan][HIR] Add support for private F90DV alias analysis to HIR (#16665)
This change adds support for alias analysis in HIR. For this to work
processing and decomposition for aliases present in loop and loop UB has
to be added. Additionally, there can be cases when one alias is subset
of another alias for the same value. To support such cases aliases in
HIR are compared also against their symbase value. For interoperability
of aliases maps data structures modified Instruction field type to use
PointrUnion one which can be either Instruction or loopopt::HLDDNode.

Addresses: CMPLRLLVM-51607