[VPlan] Add VPlan verification for VPExternalUses and LiveIn/Out lists. (#17738)
Related JIRA:
[CMPLRLLVM-45764](https://jira.devtools.intel.com/browse/CMPLRLLVM-45764)

Patch to add verification functions to the VPlan verifier and
VPExternalValues covering VPlan LiveIn/Out lists and VPExternalUses,
respectively. These checks deal with the correctness of MergeIds, i.e.
that the merge IDs are consistent with the values' places in the array,
and that there are not duplicated ExternalUses that represent the same
underlying value.

The verification for VPExternalUses is under the VPExternalValues class
to match with existing verification functions. As a necessary step to
enable unit testing, there are some structural changes to the VPlan
verifier unit tests around the namespace the test base is defined in,
and friendship added to relevant classes.