CMPLRLLVM-53121: Fix for invalidating the analyses for all inner loops during HIROprPredicate (#17984)
The HIR opt-predicate will modify the parent loop of the condition, as
well the target loop (outer-loop that will be encapsulated by hoisting).
If the parent loop has a sibling loop, and some analysis pass was
applied to it (e.g. safe reduction analysis), then modififying the
target loop may affect the result of the analysis for the sibling loop
too. Any other pass that use the modified loop will be affected and may
produce a segfault. This patch makes sure that we invalidate all the
analyses for the target loop, as well all the nested loops.

This is the same PR as #17796 but for xmain-rel.