[SYCL][Assert] Fix multi-def of assert fallback on fpga emu (#18012) (#18052)
RemoveDeviceLibAssertFallback pass will remove devicelib assert
fail definition and we will use the one defined in OCL RT for CPU
device.

But as we define the device assert faill / read funcition for both
cpu device and fpga emu device. If it only removes devicelib assert
fail definition for CPU device, it introduces multiple definition
for fpga emu.

Our solution is defining the name of assert function on OCL device with
"_opencl" suffix. And we rename the callee to
"__devicelib_assert_fail_opencl" and "__devicelib_assert_read_opencl"
for cpu device. It will fix the multiple definition issue for fpga emu.