[SYCL] Disable alignment property on annotated_ptr (#18151)
This change is only made for 2024.1 release, because when applying
compile-time `alignment` property on annotated_ptr class, it may cause
functional issue when operations like
```
annotated_ptr  p = ...;
p[i] = ...;
p++;
```
when offsetting the annotated_ptr, the alignment is no longer the same
as the annotation has specified. The fix for this issue is planned in
2024.2

Related e2e tests have been updated.