CMPLRLLVM-53121: Include fake refs during the collection in loop reversal (#18251)
The loop reversal opt checks isn't applied when there are unsafe calls.
On the other hand,
when the calls are safe and can be transformed, the transformation it
didn't modify the
fake refs, which produces an assertion in the HIR verifier. This patch
update the mem refs
collectors to include the fake refs and modify them.

This is the same PR as #17985 but for xmain-rel.