[OptReport][Driver] Update debug info usage for optimization reports (#18287)
[OptReport] Add flag to disable eventual debug info emission (#18119)

Debug information is enabled by the driver when the optimization report
is because it's required for showing source code locations in the
report. However, it shouldn't be emitted in the final binary unless the
user has explicitly passed -g (or an equivalent option). This change
adds a flag that can be used by the driver to disable debug info
emission for the final binary while keeping debug info during
optimization, which it can use in the case where -qopt-report is passed
but -g is not.

[Driver] Update opt-report usage to limit debug information (#18233)

When using -qopt-report, we enable debug information.  Add an additional
switch to make sure that information is only used for the optimization
reports and is not emitted to the final binary.

Jira: CMPLRLLVM-53856
Cherry-pick 1:
https://github.com/intel-restricted/applications.compilers.llvm-project/pull/18119
Cherry-pick 2:
https://github.com/intel-restricted/applications.compilers.llvm-project/pull/18233

Co-authored-by: Daniel Woodworth <daniel.woodworth@intel.com>