[Paropt] Ensure that Functions in MAP:FPTR are passed as kernel args. (#18079) (#18281)
Functions can be map clause operands (with FPTR modifier) for cases
like:

```llvm
    declare void @foo();
    store @foo, ptr fp

    "MAP.FPTR"(ptr %fp, ...)
```

which after optimizations can become:

```llvm
    "MAP.FPTR"(ptr @foo, ...)
```

This optimization would remove the kernel argument for `%fp`. We do
renaming of Globals to ensure that they pass through to the kernel as
arguments, but not Functions. This can cause a difference in the
number of host and device kernel args for FPTRs.

So, we need to do that renaming for Functions for FPTR maps, to
make sure they get passed as kernel args.

JIRA: CMPLRLLVM-54005
Xmain PR: #18079
Prerequisite: #18197
(cherry picked from commit 59494634d39ee95d7da638bc7aa73a2f7708d625)